requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
aiogram==3.3.0
openai==1.6.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.2
python-dotenv==1.0.0

pytest==7.4.3
pytest-asyncio==1.1.0